Enable coexistence with other libSBTL based modules
When multiple libSBTL modules are used together and they all have a
file with the same name, there are naming conflict. This could be
avoided by adding include guards, but every module would have to
have a prefix, so that the include guard has a unique name. However,
we can actually specify the path into the contrib dir which avoid
the problem as well.

Refs #000